// src/app/models/chi-tiet-san-pham.model.ts
export class ChiTietSanPhamModel{
  id: number | undefined;
  ma: string = '';
  idSanPham: number | undefined;
  tenSanPham: string = '';
  idMauSac: number | undefined;
  maMauSac: string = '';
  tenMauSac: string = '';
  idKichThuoc: number | undefined;
  tenKichThuoc: string = '';
  donGia: number = 0;
  giaBan: number = 0;
  soLuong: number = 0;
  ghiChu: string = '';
  trangThai: number = 0;
  image: string = '';
}